+++ .travis.yml
@@ -14,7 +14,7 @@ notifications:
     on_success: change
 before_install:
 - gem install cocoapods
-- pod install --project-directory=Example
+- pod install
 script:
-# - set -o pipefail && xcodebuild clean test -workspace Example/LaunchGate.xcworkspace -scheme LaunchGate-Example -destination 'platform=iOS Simulator,name=iPhone 6s' ONLY_ACTIVE_ARCH=NO | xcpretty
-- pod lib lint --allow-warnings
+- set -o pipefail && xcodebuild clean test -workspace LaunchGate.xcworkspace -scheme LaunchGate -destination 'platform=iOS Simulator,name=iPhone 6s' ONLY_ACTIVE_ARCH=NO | xcpretty
+- pod lib
