--- Example/Podfile
@@ -1,13 +1,0 @@
-source 'https://github.com/CocoaPods/Specs.git'
-use_frameworks!
-
-target 'LaunchGate_Example', :exclusive => true do
-  pod "LaunchGate", :path => "../"
-end
-
-target 'LaunchGate_Tests', :exclusive => true do
-  pod "LaunchGate", :path => "../"
-
-  pod 'Quick', '~> 0.8.0'
-  pod 'Nimble', '3.0.0'
-end
