--- Pod/Classes/LaunchGateDialogManager.swift
@@ -1,25 +1,0 @@
-//
-//  LaunchGateDialogManager.swift
-//  Pods
-//
-//  Created by Dan Trenz on 2/5/16.
-//
-//
-
-import Foundation
-
-class LaunchGateDialogManager {
-
-  func displayAlertDialog(message: String) {
-    // TBD
-  }
-
-  func displayRequiredUpdateDialog(message: String) {
-    // TBD
-  }
-
-  func displayOptionalUpdateDialog(message: String) {
-    // TBD
-  }
-
-}
