+++ Podfile
@@ -1,0 +1,7 @@
+source 'https://github.com/CocoaPods/Specs.git'
+use_frameworks!
+
+target 'LaunchGateTests', :exclusive => true do
+  pod 'Quick', '~> 0.8.0'
+  pod 'Nimble', '3.0.0'
+end
