+++ Source/LaunchGate.h
@@ -1,0 +1,19 @@
+//
+//  LaunchGate.h
+//  LaunchGate
+//
+//  Created by Dan Trenz on 2/8/16.
+//  Copyright © 2016 Dan Trenz. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for LaunchGate.
+FOUNDATION_EXPORT double LaunchGateVersionNumber;
+
+//! Project version string for LaunchGate.
+FOUNDATION_EXPORT const unsigned char LaunchGateVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <LaunchGate/PublicHeader.h>
+
+
