+++ Source/LaunchGateConfiguration.swift
@@ -1,0 +1,18 @@
+//
+//  LaunchGateConfiguration
+//  Pods
+//
+//  Created by Dan Trenz on 2/2/16.
+//
+//
+
+import Foundation
+
+
+public struct LaunchGateConfiguration {
+
+  var alert: AlertConfiguration?
+  var optionalUpdate: UpdateConfiguration?
+  var requiredUpdate: UpdateConfiguration?
+
+}
