+++ Tests/ConfigurationSpec.swift
@@ -1,0 +1,12 @@
+import Quick
+import Nimble
+
+class ConfigurationSpec: QuickSpec {
+  override func spec() {
+    
+    describe("") {
+      
+    }
+
+  }
+}
