+++ Tests/Fixtures/config-malformed-optional.json
