+++ Tests/Fixtures/config-malformed-required.json
